requirements.txt
# Core HTTP and async operations
aiohttp>=3.9.0
aiofiles>=23.2.1

# HTML parsing and web scraping
beautifulsoup4>=4.12.0
lxml>=4.9.0

# DNS and domain operations
dnspython>=2.4.0
python-whois>=0.8.0
tldextract>=5.0.0

# Validation
validators>=0.22.0

# Reporting and templating
jinja2>=3.1.2

# Terminal output
colorama>=0.4.6

# Configuration
PyYAML>=6.0.1

# Testing (dev dependencies)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code quality (dev dependencies)
black>=23.7.0
flake8>=6.1.0
bandit>=1.7.5
